requirements.txt
streamlit
torch
opencv-python
Pillow
numpy
faiss-cpu
open-clip-torch
